backend/requirements.txt
# DEVE FICAR ASSIM:
fastapi
uvicorn[standard]
pydantic
python-dotenv
supabase